+++ src/main/java/com/openrca/ecommerce/repository/AdminRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdminRepository extends JpaRepository<AdminRepository,Integer> {
+public interface AdminRepository extends JpaRepository<AdminRepository,Long> {
 
 }
